### 2026-02-25 #287 @chaitanyakasaraneni
Hey @crhea93 ! Thank you for looking at this!

The code was developed iteratively over the past several weeks as a generalization of patterns I have used across multiple production clinical ML projects at my work, where I build bioinformatics data pipelines that process genomic and clinical data.

The library was developed locally and then published to GitHub and PyPI on February 22, 2026. There is no prior public repository — this was a deliberate decision to package recurring internal patterns as a standalone open-source library rather than extracting them from proprietary codebases.

The large second commit reflects the initial public release of code that had been written locally before the repository was created. I understand this is a known pattern that raises questions about development history, and I want to be transparent about it.

Please let me know if you have any other questions before deciding whether to proceed.
