### 2026-02-26 #287 @crhea93
@chaitanyakasaraneni Thank you for the explanation. Yes, this is a pattern common in vibe-coded work; but, your comment clearly explains the case :) I'm happy to continue with this.
