+++ clinops/ingest/mimic_iii.py
@@ -1,0 +1,585 @@
+"""
+MIMIC-III clinical database loader.
+
+``MimicIIILoader`` provides the same interface as ``MimicLoader`` (MIMIC-IV)
+but targets the MIMIC-III schema, which differs in three important ways:
+
+1. **Flat directory structure** — all tables live in a single directory,
+   no ``hosp/`` or ``icu/`` subdirectories.
+2. **Uppercase column names** — ``SUBJECT_ID``, ``HADM_ID``, ``ICUSTAY_ID``,
+   ``CHARTTIME``, etc. Normalised to lowercase on load so downstream code
+   can treat MIMIC-III and MIMIC-IV DataFrames identically.
+3. **ICD-9-CM only** — no ICD-10 codes; ``icd_version`` column is always 9.
+
+This loader is intended for researchers benchmarking against the large body
+of existing MIMIC-III literature (MIMIC-Extract, DeepPatient, AIM-III, etc.)
+who want to reuse clinops preprocessing pipelines without rewriting filters.
+
+Supported tables
+----------------
+- ``chartevents``    — ICU charted observations (vitals, GCS, vents)
+- ``labevents``      — Hospital lab results
+- ``admissions``     — Hospital admission / discharge records
+- ``diagnoses_icd``  — ICD-9-CM diagnosis codes (seq_num + icd_code)
+- ``icustays``       — ICU stay metadata including LOS
+- ``prescriptions``  — Medication orders
+- ``inputevents_mv`` — MetaVision fluid inputs (CareVue: ``inputevents_cv``)
+- ``d_items``        — Item dictionary (maps itemid → label)
+- ``d_labitems``     — Lab item dictionary
+
+Examples
+--------
+>>> from clinops.ingest import MimicIIILoader
+>>>
+>>> loader = MimicIIILoader("/data/mimic-iii-clinical-database-1.4")
+>>>
+>>> charts = loader.chartevents(subject_ids=[40124, 41914])
+>>> labs   = loader.labevents(subject_ids=[40124], with_ref_range=True)
+>>> dx     = loader.diagnoses_icd(subject_ids=[40124], primary_only=True)
+
+Authors
+-------
+Chaitanya Kasaraneni
+
+References
+----------
+Johnson, A.E.W. et al. MIMIC-III, a freely accessible critical care database.
+Sci Data 3, 160035 (2016). https://doi.org/10.1038/sdata.2016.35
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from typing import Sequence
+
+import pandas as pd
+from loguru import logger
+from pydantic import BaseModel, field_validator
+
+from clinops.ingest.schema import ClinicalSchema, ColumnSpec, SchemaValidationError
+
+
+# ---------------------------------------------------------------------------
+# Column name normalisation
+# ---------------------------------------------------------------------------
+
+#: MIMIC-III uses uppercase column names; normalise to lowercase so DataFrames
+#: are compatible with MIMIC-IV loaders and downstream clinops transforms.
+def _normalise_columns(df: pd.DataFrame) -> pd.DataFrame:
+    df.columns = [c.lower() for c in df.columns]
+    return df
+
+
+# ---------------------------------------------------------------------------
+# Table registry
+# ---------------------------------------------------------------------------
+
+#: All MIMIC-III tables live in a flat directory (no module subdirectories).
+_TABLE_STEMS: dict[str, str] = {
+    "chartevents":    "CHARTEVENTS",
+    "labevents":      "LABEVENTS",
+    "admissions":     "ADMISSIONS",
[…]
