+++ tests/test_ingest/test_mimic_iii.py
@@ -1,0 +1,412 @@
+"""
+Tests for MimicIIILoader.
+
+Uses minimal uppercase CSV fixtures in a flat tmp directory — no real
+MIMIC-III data needed. Fixtures mirror the actual MIMIC-III column names
+(uppercase) to verify the normalisation logic.
+"""
+
+from __future__ import annotations
+
+import pandas as pd
+import pytest
+
+from clinops.ingest.mimic_iii import MimicIIILoader
+
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+def _write_csvs(tmp_path):
+    """Write minimal MIMIC-III-style CSV files (uppercase columns, flat dir)."""
+
+    # CHARTEVENTS — 3 rows, 2 patients, uppercase columns
+    (tmp_path / "CHARTEVENTS.csv").write_text(
+        "SUBJECT_ID,HADM_ID,ICUSTAY_ID,ITEMID,CHARTTIME,VALUENUM,VALUEUOM,ERROR\n"
+        "1,100,1000,211,2150-01-01 06:00:00,72.0,bpm,\n"
+        "1,100,1000,211,2150-01-01 12:00:00,75.0,bpm,\n"
+        "2,200,2000,211,2150-01-02 06:00:00,80.0,bpm,\n"
+    )
+
+    # LABEVENTS — 2 rows with ref range columns
+    (tmp_path / "LABEVENTS.csv").write_text(
+        "SUBJECT_ID,HADM_ID,ITEMID,CHARTTIME,VALUE,VALUENUM,VALUEUOM,"
+        "REF_RANGE_LOWER,REF_RANGE_UPPER,FLAG\n"
+        "1,100,50912,2150-01-01 08:00:00,1.1,1.1,mg/dL,0.5,1.5,\n"
+        "2,200,50912,2150-01-02 08:00:00,0.9,0.9,mg/dL,0.5,1.5,\n"
+    )
+
+    # ADMISSIONS
+    (tmp_path / "ADMISSIONS.csv").write_text(
+        "SUBJECT_ID,HADM_ID,ADMITTIME,DISCHTIME,DEATHTIME,ADMISSION_TYPE,"
+        "ADMISSION_LOCATION,DISCHARGE_LOCATION,INSURANCE,HOSPITAL_EXPIRE_FLAG\n"
+        "1,100,2150-01-01 00:00:00,2150-01-05 00:00:00,,EMERGENCY,"
+        "EMERGENCY ROOM,HOME,Medicare,0\n"
+        "2,200,2150-01-02 00:00:00,2150-01-07 00:00:00,,ELECTIVE,"
+        "PHYSICIAN REFERRAL,HOME,Private,0\n"
+    )
+
+    # DIAGNOSES_ICD — ICD-9 only, mix of numeric and alphanumeric codes
+    (tmp_path / "DIAGNOSES_ICD.csv").write_text(
+        "SUBJECT_ID,HADM_ID,SEQ_NUM,ICD9_CODE\n"
+        "1,100,1,4280\n"
+        "1,100,2,V053\n"
+        "2,200,1,41401\n"
+        "2,200,2,5849\n"
+    )
+
+    # ICUSTAYS
+    (tmp_path / "ICUSTAYS.csv").write_text(
+        "SUBJECT_ID,HADM_ID,ICUSTAY_ID,FIRST_CAREUNIT,LAST_CAREUNIT,"
+        "INTIME,OUTTIME,LOS\n"
[…]
