### 2026-03-01 bcf1eb1
tests: update test coverage and minor bug fix in temporal
