+++ tests/test_temporal/test_temporal.py
@@ -85,6 +85,172 @@ class TestTemporalWindower:
         result = windower.fit_transform(df, "subject_id", "charttime", ["heart_rate"])
         assert result["subject_id"].nunique() == 5
 
+    def test_long_format_pivot_produces_wide_output(self):
+        base = datetime(2023, 1, 1)
+        rows = []
+        for h in range(12):
+            t = base + timedelta(hours=h)
+            rows.append({"subject_id": 1, "charttime": t, "itemid": 220045, "valuenum": 70.0 + h})
+            rows.append({"subject_id": 1, "charttime": t, "itemid": 220277, "valuenum": 97.0})
+        df = pd.DataFrame(rows)
+        windower = TemporalWindower(window_hours=6, step_hours=6)
+        result = windower.fit_transform(
+            df, id_col="subject_id", time_col="charttime",
+            item_col="itemid", value_col="valuenum",
+        )
+        assert len(result) > 0
+        col_names = [str(c) for c in result.columns]
+        assert "220045" in col_names or 220045 in result.columns
+
+    def test_pivot_columns_name_is_none(self):
+        base = datetime(2023, 1, 1)
+        rows = [{"subject_id": 1, "charttime": base + timedelta(hours=h),
+                 "itemid": 220045, "valuenum": 70.0} for h in range(6)]
+        df = pd.DataFrame(rows)
+        windower = TemporalWindower(window_hours=6, step_hours=6)
+        result = windower.fit_transform(
+            df, id_col="subject_id", time_col="charttime",
+            item_col="itemid", value_col="valuenum",
+        )
+        assert result.columns.name is None
+
+    def test_feature_cols_auto_detected(self):
+        base = datetime(2023, 1, 1)
+        df = pd.DataFrame({
+            "subject_id": [1] * 6,
+            "charttime": [base + timedelta(hours=h) for h in range(6)],
+            "heart_rate": [70.0] * 6,
+            "spo2": [97.0] * 6,
+        })
+        windower = TemporalWindower(window_hours=6, step_hours=6)
+        result = windower.fit_transform(df, id_col="subject_id", time_col="charttime")
+        assert "heart_rate" in result.columns
+        assert "spo2" in result.columns
+
+    def test_missing_feature_col_produces_nan(self):
+        base = datetime(2023, 1, 1)
+        df = pd.DataFrame({
+            "subject_id": [1] * 6,
+            "charttime": [base + timedelta(hours=h) for h in range(6)],
+            "heart_rate": [70.0] * 6,
+        })
+        windower = TemporalWindower(window_hours=6, step_hours=6)
+        result = windower.fit_transform(
+            df, id_col="subject_id", time_col="charttime",
+            feature_cols=["heart_rate", "nonexistent"],
+        )
+        assert "nonexistent" in result.columns
+        assert result["nonexistent"].isna().all()
+
+    def test_all_nan_feature_within_window_gives_nan(self):
+        base = datetime(2023, 1, 1)
+        df = pd.DataFrame({
+            "subject_id": [1, 1, 1],
+            "charttime": [base + timedelta(hours=h) for h in range(3)],
+            "heart_rate": [np.nan, np.nan, np.nan],
+        })
+        windower = TemporalWindower(window_hours=6, step_hours=6, min_observations=1)
+        result = windower.fit_transform(
[…]
