+++ tests/test_ingest/test_ingest.py
@@ -184,3 +184,234 @@ class TestFHIRLoader:
     def test_source_not_found(self):
         with pytest.raises(FileNotFoundError):
             FHIRLoader("/nonexistent/fhir")
+
+    def test_observations_filter_by_loinc_code(self, tmp_path):
+        import json
+
+        bundle = {
+            "resourceType": "Bundle",
+            "entry": [
+                {
+                    "resource": {
+                        "resourceType": "Observation",
+                        "id": "o1",
+                        "subject": {"reference": "Patient/p1"},
+                        "code": {"coding": [{"system": "http://loinc.org", "code": "8867-4"}]},
+                        "valueQuantity": {"value": 72.0, "unit": "/min"},
+                        "effectiveDateTime": "2023-01-01T08:00:00Z",
+                        "status": "final",
+                    }
+                },
+                {
+                    "resource": {
+                        "resourceType": "Observation",
+                        "id": "o2",
+                        "subject": {"reference": "Patient/p1"},
+                        "code": {"coding": [{"system": "http://loinc.org", "code": "59408-5"}]},
+                        "valueQuantity": {"value": 98.0, "unit": "%"},
+                        "effectiveDateTime": "2023-01-01T08:00:00Z",
+                        "status": "final",
+                    }
+                },
+            ],
+        }
+        f = tmp_path / "obs.json"
+        f.write_text(json.dumps(bundle))
+        loader = FHIRLoader(f)
+        df = loader.observations(loinc_codes=["8867-4"])
+        assert len(df) == 1
+        assert df.iloc[0]["loinc_code"] == "8867-4"
+
+    def test_conditions_basic(self, tmp_path):
+        import json
+
+        bundle = {
+            "resourceType": "Bundle",
+            "entry": [
+                {
+                    "resource": {
+                        "resourceType": "Condition",
+                        "id": "c1",
+                        "subject": {"reference": "Patient/p1"},
+                        "code": {
+                            "coding": [
+                                {
+                                    "system": "http://snomed.info/sct",
+                                    "code": "73211009",
+                                    "display": "Diabetes mellitus",
+                                }
+                            ]
+                        },
+                        "clinicalStatus": {"coding": [{"code": "active"}]},
+                        "onsetDateTime": "2020-03-01",
+                    }
+                }
+            ],
+        }
+        f = tmp_path / "conditions.json"
+        f.write_text(json.dumps(bundle))
+        loader = FHIRLoader(f)
+        df = loader.conditions()
+        assert len(df) == 1
+        assert df.iloc[0]["condition_id"] == "c1"
+        assert df.iloc[0]["patient_id"] == "p1"
+        assert df.iloc[0]["code"] == "73211009"
+        assert df.iloc[0]["system"] == "http://snomed.info/sct"
+        assert df.iloc[0]["display"] == "Diabetes mellitus"
+        assert df.iloc[0]["clinical_status"] == "active"
+        assert df.iloc[0]["onset"] == "2020-03-01"
+
+    def test_conditions_no_codings_gives_none(self, tmp_path):
+        import json
+
+        bundle = {
+            "resourceType": "Bundle",
+            "entry": [
+                {
+                    "resource": {
+                        "resourceType": "Condition",
+                        "id": "c2",
+                        "subject": {"reference": "Patient/p2"},
[…]
