+++ clinops/ingest/__init__.py
@@ -12,10 +12,12 @@ from clinops.ingest.fhir import FHIRLoader
 from clinops.ingest.flat import FlatFileLoader
 from clinops.ingest.mimic import MimicLoader
 from clinops.ingest.mimic_tables import MimicTableLoader
+from clinops.ingest.mimic_iii import MimicIIILoader
 from clinops.ingest.schema import ClinicalSchema, ColumnSpec, SchemaValidationError
 
 __all__ = [
     "MimicLoader",
+    "MimicIIILoader",
     "MimicTableLoader",
     "FHIRLoader",
     "FlatFileLoader",
