+++ tests/test_ingest/test_mimic_iii.py
@@ -13,7 +13,6 @@ import pytest
 
 from clinops.ingest.mimic_iii import MimicIIILoader
 
-
 # ---------------------------------------------------------------------------
 # Fixtures
 # ---------------------------------------------------------------------------
