+++ clinops/ingest/__init__.py
@@ -11,8 +11,8 @@ Supported sources:
 from clinops.ingest.fhir import FHIRLoader
 from clinops.ingest.flat import FlatFileLoader
 from clinops.ingest.mimic import MimicLoader
-from clinops.ingest.mimic_tables import MimicTableLoader
 from clinops.ingest.mimic_iii import MimicIIILoader
+from clinops.ingest.mimic_tables import MimicTableLoader
 from clinops.ingest.schema import ClinicalSchema, ColumnSpec, SchemaValidationError
 
 __all__ = [
