### 2026-03-02 e49c351
fix: update tests as per suggestions
