+++ clinops/ingest/mimic_iii.py
@@ -20,7 +20,8 @@ Supported tables
 - ``chartevents``    — ICU charted observations (vitals, GCS, vents)
 - ``labevents``      — Hospital lab results
 - ``admissions``     — Hospital admission / discharge records
-- ``diagnoses_icd``  — ICD-9-CM diagnosis codes (seq_num + icd_code)
+- ``patients``       — Patient-level demographics
+- ``diagnoses_icd``  — ICD-9-CM diagnosis codes (seq_num + icd9_code)
 - ``icustays``       — ICU stay metadata including LOS
 - ``prescriptions``  — Medication orders
 - ``inputevents_mv`` — MetaVision fluid inputs (CareVue: ``inputevents_cv``)
@@ -49,7 +50,6 @@ Sci Data 3, 160035 (2016). https://doi.org/10.1038/sdata.2016.35
 
 from __future__ import annotations
 
-import re
 from collections.abc import Sequence
 from pathlib import Path
 
@@ -76,38 +76,30 @@ def _normalise_columns(df: pd.DataFrame) -> pd.DataFrame:
 
 #: All MIMIC-III tables live in a flat directory (no module subdirectories).
 _TABLE_STEMS: dict[str, str] = {
-    "chartevents":    "CHARTEVENTS",
-    "labevents":      "LABEVENTS",
-    "admissions":     "ADMISSIONS",
-    "diagnoses_icd":  "DIAGNOSES_ICD",
-    "icustays":       "ICUSTAYS",
-    "prescriptions":  "PRESCRIPTIONS",
+    "chartevents": "CHARTEVENTS",
+    "labevents": "LABEVENTS",
+    "admissions": "ADMISSIONS",
+    "diagnoses_icd": "DIAGNOSES_ICD",
+    "icustays": "ICUSTAYS",
+    "prescriptions": "PRESCRIPTIONS",
     "inputevents_mv": "INPUTEVENTS_MV",
     "inputevents_cv": "INPUTEVENTS_CV",
-    "d_items":        "D_ITEMS",
-    "d_labitems":     "D_LABITEMS",
-    "patients":       "PATIENTS",
+    "d_items": "D_ITEMS",
+    "d_labitems": "D_LABITEMS",
+    "patients": "PATIENTS",
 }
 
 #: Required columns per table (post-normalisation, lowercase).
 _REQUIRED_COLS: dict[str, list[str]] = {
-    "chartevents":    ["subject_id", "hadm_id", "icustay_id", "itemid",
-                       "charttime", "valuenum"],
-    "labevents":      ["subject_id", "hadm_id", "itemid", "charttime",
-                       "valuenum", "valueuom"],
-    "admissions":     ["subject_id", "hadm_id", "admittime", "dischtime",
-                       "admission_type"],
-    "diagnoses_icd":  ["subject_id", "hadm_id", "seq_num", "icd9_code"],
-    "icustays":       ["subject_id", "hadm_id", "icustay_id", "intime",
-                       "outtime", "los"],
-    "prescriptions":  ["subject_id", "hadm_id", "startdate", "drug"],
-    "inputevents_mv": ["subject_id", "hadm_id", "icustay_id", "itemid",
-                       "starttime", "amount", "amountuom"],
-    "inputevents_cv": ["subject_id", "hadm_id", "icustay_id", "itemid",
-                       "charttime", "amount", "amountuom"],
-    "d_items":        ["itemid", "label"],
-    "d_labitems":     ["itemid", "label"],
-    "patients":       ["subject_id", "gender", "dob"],
+    "chartevents": ["subject_id", "hadm_id", "icustay_id", "itemid", "charttime", "valuenum"],
[…]
