### 2026-03-02 24b0cc2
fix: use UUID-based sentinel column with try/finally cleanup in imputation.py
Co-authored-by: chaitanyakasaraneni <24287691+chaitanyakasaraneni@users.noreply.github.com>
