### 2026-03-02 c4f39de
use UUID based sentinel
