### 2026-04-08 2f48219
fix: prevent ffill/bfill from crossing entity boundaries in Imputer
When id_col is set, apply fill and gap masking within each entity group
rather than globally, so values from one patient/admission cannot
propagate into another. Also exclude id_col from numeric_cols to avoid
treating it as a feature.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
