### 2026-04-08 3aae4b5
style: apply ruff formatting
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
