+++ tests/test_ingest/test_mimic_iii.py
@@ -17,6 +17,7 @@ from clinops.ingest.mimic_iii import MimicIIILoader
 # Fixtures
 # ---------------------------------------------------------------------------
 
+
 def _write_csvs(tmp_path):
     """Write minimal MIMIC-III-style CSV files (uppercase columns, flat dir)."""
 
@@ -126,6 +127,7 @@ def loader(mimic3_dir):
 # Column normalisation
 # ---------------------------------------------------------------------------
 
+
 class TestColumnNormalisation:
     def test_chartevents_columns_lowercase(self, loader):
         df = loader.chartevents()
@@ -143,6 +145,7 @@ class TestColumnNormalisation:
 # ---------------------------------------------------------------------------
 # chartevents
 # ---------------------------------------------------------------------------
+
 
 class TestChartevents:
     def test_loads_all_rows(self, loader):
@@ -183,14 +186,14 @@ class TestChartevents:
 
     def test_required_columns_present(self, loader):
         df = loader.chartevents()
-        for col in ["subject_id", "hadm_id", "icustay_id", "itemid",
-                    "charttime", "valuenum"]:
+        for col in ["subject_id", "hadm_id", "icustay_id", "itemid", "charttime", "valuenum"]:
             assert col in df.columns
 
 
 # ---------------------------------------------------------------------------
 # labevents
 # ---------------------------------------------------------------------------
+
 
 class TestLabevents:
     def test_loads_all_rows(self, loader):
@@ -218,6 +221,7 @@ class TestLabevents:
 # admissions
 # ---------------------------------------------------------------------------
 
+
 class TestAdmissions:
     def test_loads_all_rows(self, loader):
         assert len(loader.admissions()) == 2
@@ -227,14 +231,13 @@ class TestAdmissions:
         assert len(df) == 1
 
     def test_admittime_is_datetime(self, loader):
-        assert pd.api.types.is_datetime64_any_dtype(
-            loader.admissions()["admittime"]
-        )
+        assert pd.api.types.is_datetime64_any_dtype(loader.admissions()["admittime"])
 
 
 # ---------------------------------------------------------------------------
 # diagnoses_icd
 # ---------------------------------------------------------------------------
+
 
 class TestDiagnosesIcd:
     def test_loads_all_rows(self, loader):
@@ -282,6 +285,7 @@ class TestDiagnosesIcd:
 # icustays
 # ---------------------------------------------------------------------------
 
+
 class TestIcustays:
     def test_loads_all_rows(self, loader):
         assert len(loader.icustays()) == 2
@@ -309,6 +313,7 @@ class TestIcustays:
 # prescriptions
 # ---------------------------------------------------------------------------
 
+
 class TestPrescriptions:
     def test_loads_all_rows(self, loader):
         assert len(loader.prescriptions()) == 3
@@ -325,14 +330,13 @@ class TestPrescriptions:
         assert len(loader.prescriptions(drugs=["heparin"])) == 1
 
     def test_startdate_is_datetime(self, loader):
-        assert pd.api.types.is_datetime64_any_dtype(
-            loader.prescriptions()["startdate"]
-        )
+        assert pd.api.types.is_datetime64_any_dtype(loader.prescriptions()["startdate"])
 
 
 # ---------------------------------------------------------------------------
 # inputevents
 # ---------------------------------------------------------------------------
+
 
 class TestInputevents:
     def test_mv_loads(self, loader):
@@ -396,6 +400,7 @@ class TestInputevents:
[…]
