+++ clinops/ingest/mimic_iii.py
@@ -63,6 +63,7 @@ from clinops.ingest.schema import SchemaValidationError
 # Column name normalisation
 # ---------------------------------------------------------------------------
 
+
 #: MIMIC-III uses uppercase column names; normalise to lowercase so DataFrames
 #: are compatible with MIMIC-IV loaders and downstream clinops transforms.
 def _normalise_columns(df: pd.DataFrame) -> pd.DataFrame:
@@ -98,19 +99,24 @@ _REQUIRED_COLS: dict[str, list[str]] = {
     "icustays": ["subject_id", "hadm_id", "icustay_id", "intime", "outtime", "los"],
     "prescriptions": ["subject_id", "hadm_id", "startdate", "drug"],
     "inputevents_mv": [
-        "subject_id", "hadm_id", "icustay_id", "itemid", "starttime", "amount", "amountuom"
+        "subject_id",
+        "hadm_id",
+        "icustay_id",
+        "itemid",
+        "starttime",
+        "amount",
+        "amountuom",
     ],
 }
 
 #: Datetime columns per table (post-normalisation, lowercase).
 _DATETIME_COLS: dict[str, list[str]] = {
-    "chartevents":    ["charttime"],
-    "labevents":      ["charttime"],
-    "admissions":     ["admittime", "dischtime", "deathtime", "edregtime",
-                       "edouttime"],
-    "diagnoses_icd":  [],
-    "icustays":       ["intime", "outtime"],
-    "prescriptions":  ["startdate", "enddate"],
+    "chartevents": ["charttime"],
+    "labevents": ["charttime"],
+    "admissions": ["admittime", "dischtime", "deathtime", "edregtime", "edouttime"],
+    "diagnoses_icd": [],
+    "icustays": ["intime", "outtime"],
+    "prescriptions": ["startdate", "enddate"],
     "inputevents_mv": ["starttime", "endtime"],
     "inputevents_cv": ["charttime"],
 }
@@ -119,6 +125,7 @@ _DATETIME_COLS: dict[str, list[str]] = {
 # ---------------------------------------------------------------------------
 # Config
 # ---------------------------------------------------------------------------
+
 
 class MimicIIIConfig(BaseModel):
     """Validated configuration for MimicIIILoader."""
@@ -138,6 +145,7 @@ class MimicIIIConfig(BaseModel):
 # ---------------------------------------------------------------------------
 # MimicIIILoader
 # ---------------------------------------------------------------------------
+
 
 class MimicIIILoader:
     """
@@ -185,9 +193,7 @@ class MimicIIILoader:
             strict_validation=strict_validation,
             chunk_size=chunk_size,
         )
-        logger.info(
-            f"MimicIIILoader initialised — path={self._cfg.mimic_path}"
-        )
+        logger.info(f"MimicIIILoader initialised — path={self._cfg.mimic_path}")
 
     # ------------------------------------------------------------------
     # Public table accessors
@@ -230,7 +236,9 @@ class MimicIIILoader:
         """
         df = self._load("chartevents")
         return self._filter(
-            df, subject_ids, hadm_ids,
+            df,
+            subject_ids,
+            hadm_ids,
             icustay_ids=icustay_ids,
             item_ids=item_ids,
             start_time=start_time,
@@ -266,15 +274,16 @@ class MimicIIILoader:
         """
         df = self._load("labevents")
         df = self._filter(
-            df, subject_ids, hadm_ids,
+            df,
+            subject_ids,
+            hadm_ids,
             item_ids=item_ids,
             start_time=start_time,
             end_time=end_time,
             time_col="charttime",
         )
         if not with_ref_range:
-            drop = [c for c in ["ref_range_lower", "ref_range_upper"]
-                    if c in df.columns]
[…]
