### 2026-04-08 c8ab3db
fix: resolve mypy type errors in imputation and mimic_iii
- Assert time_col is not None in _fill_with_gap_mask to narrow str|None
- Annotate read_kwargs as dict[str, object] to satisfy type-arg check

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
