### 2026-04-08 ff3764d
fix: pass time_col explicitly to _fill_with_gap_mask instead of asserting
Assert is stripped under -O and signals the wrong intent. Passing time_col
as a str parameter lets mypy narrow the type at the call site, where the
caller already guarantees it is non-None.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
