### 2026-04-08 8369195
docs: add MkDocs site for ReadTheDocs
- .readthedocs.yaml — RTD build config (MkDocs + Python 3.12, installs dev extras)
- mkdocs.yml — Material theme, mkdocstrings for auto API docs, nav structure
- docs/index.md — landing page
- docs/getting-started.md — install + end-to-end walkthrough
- docs/guide/{ingest,preprocess,temporal,split}.md — user guides per module
- docs/api/{ingest,preprocess,temporal,split}.md — API reference via mkdocstrings
- docs/contributing.md — setup, test, lint, and PR checklist

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
