+++ docs/api/preprocess.md
@@ -1,0 +1,7 @@
+# clinops.preprocess
+
+::: clinops.preprocess.outliers.ClinicalOutlierClipper
+
+::: clinops.preprocess.units.UnitNormalizer
+
+::: clinops.preprocess.icd.ICDMapper
