+++ docs/api/split.md
@@ -1,0 +1,9 @@
+# clinops.split
+
+::: clinops.split.splitters.TemporalSplitter
+
+::: clinops.split.splitters.PatientSplitter
+
+::: clinops.split.splitters.StratifiedPatientSplitter
+
+::: clinops.split.splitters.SplitResult
