+++ docs/api/ingest.md
@@ -1,0 +1,17 @@
+# clinops.ingest
+
+::: clinops.ingest.mimic_tables.MimicTableLoader
+
+::: clinops.ingest.mimic.MimicLoader
+
+::: clinops.ingest.mimic_iii.MimicIIILoader
+
+::: clinops.ingest.fhir.FHIRLoader
+
+::: clinops.ingest.flat.FlatFileLoader
+
+::: clinops.ingest.schema.ClinicalSchema
+
+::: clinops.ingest.schema.ColumnSpec
+
+::: clinops.ingest.schema.SchemaValidationError

[…]
