### 2026-04-08 e3e6ef4
feat: add clinops.monitor and clinops.orchestrate modules
clinops.monitor
- DistributionDriftDetector: PSI + KS test per column, fit/detect pattern,
  DriftSeverity enum (LOW/MEDIUM/HIGH), DriftReport with to_dataframe() and summary()
- DataQualityChecker: null rates, schema drift (added/removed/retyped columns),
  row count anomalies, required column enforcement, QualityReport with passed/errors/warnings
- 30 tests covering stable vs drifted distributions, schema drift detection, quality issues

clinops.orchestrate
- GCSPipelineStore / S3PipelineStore: upload/download DataFrames (Parquet or C
[…]
