+++ clinops/orchestrate/store.py
@@ -1,0 +1,364 @@
+"""
+Cloud storage for clinops pipeline artifacts.
+
+Provides a consistent upload/download interface for persisting DataFrames
+and arbitrary files to GCS or S3 between pipeline steps. Designed for
+scheduled clinical data engineering workflows where intermediate results
+need to be checkpointed across runs or shared between services.
+
+Both stores use Parquet by default for efficient storage of wide clinical
+DataFrames. CSV is also supported for interoperability with downstream
+tools that cannot read Parquet.
+
+Optional dependencies
+---------------------
+GCS requires ``pip install clinops[gcp]`` (google-cloud-storage).
+S3 requires ``pip install clinops[aws]`` (boto3).
+"""
+
+from __future__ import annotations
+
+import io
+from enum import StrEnum
+from typing import Any
+
+import pandas as pd
+from loguru import logger
+
+
+class StorageFormat(StrEnum):
+    """Serialization format for pipeline artifacts."""
+
+    PARQUET = "parquet"
+    CSV = "csv"
+
+
+class GCSPipelineStore:
+    """
+    Upload and download clinops pipeline artifacts to Google Cloud Storage.
+
+    Requires ``pip install clinops[gcp]``.
+
+    Parameters
+    ----------
+    bucket:
+        GCS bucket name (without ``gs://`` prefix).
+    prefix:
+        Optional path prefix applied to all artifact names. Use to
+        namespace artifacts by pipeline or environment, e.g.
+        ``"clinops/prod/v2"``.
+    format:
+        Serialization format. Default ``StorageFormat.PARQUET``.
+    credentials_path:
+        Path to a service account JSON key file. If None, falls back to
+        Application Default Credentials (``GOOGLE_APPLICATION_CREDENTIALS``
+        env var or gcloud CLI credentials).
+
+    Examples
+    --------
+    >>> store = GCSPipelineStore("my-clinical-bucket", prefix="clinops/prod")
+    >>> store.upload(windows_df, "features/windows_2024_01")
+    'gs://my-clinical-bucket/clinops/prod/features/windows_2024_01.parquet'
+
+    >>> df = store.download("features/windows_2024_01")
+    """
+
+    def __init__(
+        self,
+        bucket: str,
+        prefix: str = "",
+        format: StorageFormat = StorageFormat.PARQUET,
+        credentials_path: str | None = None,
+    ) -> None:
+        self.bucket = bucket
+        self.prefix = prefix.rstrip("/")
+        self.format = format
+        self.credentials_path = credentials_path
+        self._client: Any = None
+
+    def upload(self, df: pd.DataFrame, name: str) -> str:
+        """
+        Serialize ``df`` and upload it to GCS.
+
+        Parameters
+        ----------
+        df:
+            DataFrame to upload.
+        name:
+            Artifact name (without extension). Forward slashes create
+            virtual directories, e.g. ``"features/windows_2024_01"``.
+
+        Returns
+        -------
+        str
+            Full GCS URI of the uploaded object (``gs://bucket/path``).
+        """
+        client = self._get_client()
+        blob_path = self._blob_path(name)
+        buffer = self._serialize(df)
+
+        bucket_obj = client.bucket(self.bucket)
+        blob = bucket_obj.blob(blob_path)
+        blob.upload_from_file(buffer, content_type=self._content_type())
+
+        uri = f"gs://{self.bucket}/{blob_path}"
+        logger.info(f"GCSPipelineStore: uploaded {len(df):,} rows → {uri}")
+        return uri
+
+    def download(self, name: str) -> pd.DataFrame:
+        """
+        Download and deserialize a DataFrame artifact from GCS.
+
+        Parameters
+        ----------
+        name:
[…]
