+++ clinops/orchestrate/stepfunctions.py
@@ -1,0 +1,360 @@
+"""
+AWS Step Functions pipeline builder for clinops workflows.
+
+Step Functions is a natural fit for clinical data pipelines: each stage
+(ingest → preprocess → window → split → train) is an independent Lambda or
+ECS task, failures can be retried automatically, and the execution history
+provides a full audit trail — often required for clinical research data
+governance.
+
+This module provides a lightweight builder that constructs a sequential
+Step Functions state machine definition from a list of :class:`PipelineStep`
+objects. The state machine can be inspected as a dict (Amazon States Language),
+deployed to AWS, or executed with a JSON payload.
+
+Requires ``pip install clinops[aws]``.
+"""
+
+from __future__ import annotations
+
+import json
+import time
+from dataclasses import dataclass, field
+from typing import Any
+
+from loguru import logger
+
+
+@dataclass
+class PipelineStep:
+    """
+    A single task in a Step Functions state machine.
+
+    Parameters
+    ----------
+    name:
+        Human-readable step name. Used as the state name in the ASL
+        definition — must be unique within a pipeline.
+    resource:
+        ARN of the Lambda function, ECS task, or Step Functions activity
+        that executes this step. See AWS docs for supported resource ARN
+        formats.
+    parameters:
+        Static input parameters merged into the state's ``Parameters``
+        block. Dynamic input from the previous step's output is also
+        available via ``$$.Execution.Input`` or ``States.JsonMerge``.
+    retry_attempts:
+        Number of times to retry on ``Lambda.ServiceException``,
+        ``Lambda.TooManyRequestsException``, or ``States.TaskFailed``.
+        Default 3.
+    timeout_seconds:
+        Maximum execution time for this step in seconds. Default 3600 (1h).
+    comment:
+        Optional human-readable description embedded in the ASL state.
+
+    Examples
+    --------
+    >>> step = PipelineStep(
+    ...     name="Preprocess",
+    ...     resource="arn:aws:lambda:us-east-1:123456789:function:clinops-preprocess",
+    ...     parameters={"max_null_rate": 0.5},
+    ...     retry_attempts=2,
+    ... )
+    """
+
+    name: str
+    resource: str
+    parameters: dict[str, Any] = field(default_factory=dict)
+    retry_attempts: int = 3
+    timeout_seconds: int = 3600
+    comment: str = ""
+
+    def to_asl_state(self, next_state: str | None) -> dict[str, Any]:
+        """
+        Render this step as an Amazon States Language Task state dict.
+
+        Parameters
+        ----------
+        next_state:
+            Name of the following state, or None if this is the terminal state.
+
+        Returns
+        -------
+        dict[str, Any]
+            ASL Task state definition.
+        """
+        state: dict[str, Any] = {
+            "Type": "Task",
+            "Resource": self.resource,
+            "TimeoutSeconds": self.timeout_seconds,
+            "Retry": [
+                {
+                    "ErrorEquals": [
+                        "Lambda.ServiceException",
+                        "Lambda.TooManyRequestsException",
+                        "States.TaskFailed",
+                    ],
+                    "IntervalSeconds": 2,
+                    "MaxAttempts": self.retry_attempts,
+                    "BackoffRate": 2.0,
+                }
+            ],
+        }
+        if self.comment:
+            state["Comment"] = self.comment
+        if self.parameters:
+            state["Parameters"] = self.parameters
+        if next_state is not None:
+            state["Next"] = next_state
+        else:
+            state["End"] = True
+        return state
+
+
+class StepFunctionsPipeline:
+    """
+    Build, deploy, and execute an AWS Step Functions state machine.
+
+    Constructs a sequential pipeline from :class:`PipelineStep` objects.
+    The generated state machine runs steps in the order they were added,
[…]
