### 2026-04-08 21f104e
fix: use bare type: ignore for service_account call in GCSPipelineStore
Specific-code ignore[no-untyped-call] triggers unused-ignore under strict
mode when the google stubs are absent. Bare type: ignore suppresses the
error without risking unused-ignore.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
