### 2026-04-08 a555fc3
fix: remove invalid pydantic[mypy] extra from dev dependencies
pydantic v2 ships the mypy plugin in the main package — the [mypy] extra
no longer exists and causes pip to warn (and potentially fail in strict
CI environments) on install.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
