+++ pyproject.toml
@@ -42,7 +42,7 @@ aws  = ["boto3>=1.34"]
 dev  = [
     "pytest>=7.4", "pytest-cov>=4.1", "pytest-mock>=3.12",
     "black==26.1.0", "ruff==0.15.2", "mypy==1.19.1", "pandas-stubs>=2.1",
-    "pydantic[mypy]>=2.5",
+    "pydantic>=2.5",
     "mkdocs>=1.5", "mkdocs-material>=9.5", "mkdocstrings[python]>=0.24",
 ]
 
