### 2026-04-08 ca276d2
fix: mypy issues with unused
