### 2026-04-08 4c8a271
add: pre-commit checks to handle formatting errors
