+++ pyproject.toml
@@ -44,6 +44,7 @@ dev  = [
     "black==26.1.0", "ruff==0.15.2", "mypy==1.19.1", "pandas-stubs>=2.1",
     "pydantic>=2.5",
     "mkdocs>=1.5", "mkdocs-material>=9.5", "mkdocstrings[python]>=0.24",
+    "pre-commit>=3.7",
 ]
 
 [project.urls]
