### 2026-04-10 #287 @crhea93
@chaitanyakasaraneni Wow I totally missed that! Where is this noted in the repo? I suggest having it in the about section (see attached for an example) and/or first thing in the README under the badges
