### 2026-04-10 658bcfd
docs: add walk-through example notebooks and examples doc page
