### 2026-04-10 1c7ffbb
docs: add ReadTheDocs badge and explicit docs/PyPI links to README
