### 2026-04-10 #287 @chaitanyakasaraneni
Thanks for the flag @crhea93! I added the URL in both places (in the About section and under the badges in the README, too).
