### 2026-06-09 e58deed
feat(ingest): add eICU-CRD adapter with leakage-guarded feature/sequence builder and patient-grouped splitter (v0.3.0)
