+++ tests/test_ingest/test_eicu.py
@@ -1,0 +1,539 @@
+"""
+Tests for the eICU-CRD adapter (clinops.ingest.eicu).
+
+All fixtures are SYNTHETIC — eICU is credentialed PhysioNet data and must never
+be committed. The synthetic tables reproduce the real eICU schema (column names,
+offset-based time, free-text labs, string age) so the loader logic is exercised
+end-to-end without any real data.
+
+Edge cases covered (per the adapter spec):
+- ``age = '> 89'``  → parsed to 90, patient retained as adult
+- a patient with two stays (``uniquepid`` repeats) → grouped split keeps them together
+- a stay with missing GCS → graceful NaN, no crash
+- an unmapped ``labname`` → logged warning, row ignored (not an error)
+- a stay above the missingness threshold → excluded from the cohort
+- the leakage guard → label-only variables never reach the feature matrix
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from clinops.ingest import EicuCohortConfig, EicuLoader, EicuTableLoader
+from clinops.ingest.eicu import (
+    FEATURE_COLUMNS,
+    LEAKAGE_FORBIDDEN,
+    ORGAN_LABELS,
+    parse_eicu_age,
+)
+from clinops.ingest.mappings.eicu_lab_map import canonical_lab_features, map_labname
+from clinops.ingest.schema import LeakageError
+from clinops.split import GroupedPatientSplitter
+
+# ---------------------------------------------------------------------------
+# Small cohort config so sequences are exercised without large fixtures.
+# ---------------------------------------------------------------------------
+
+TEST_CFG = EicuCohortConfig(
+    min_age=18,
+    min_los_hours=4,
+    max_icu_hours=8,
+    observation_hours=2,
+    prediction_hours=1,
+    missingness_threshold=0.50,
+    first_stay_only=True,
+)
+
+# Stays included by the cohort filters (adults, first stay, LOS >= 4h, low miss).
+_GOOD_STAYS = [101, 301, 401]  # P1 (normal), P3 (age > 89), P4 (missing GCS)
+
+
+def _vital_rows(stay: int, *, low_map: bool = False) -> pd.DataFrame:
+    """Dense ~30-min vitalPeriodic rows over 8h for one stay."""
+    offsets = list(range(0, 8 * 60, 30))
+    n = len(offsets)
+    rng = np.random.default_rng(stay)
+    return pd.DataFrame(
+        {
+            "vitalperiodicid": [stay * 1000 + i for i in range(n)],
+            "patientunitstayid": stay,
+            "observationoffset": offsets,
+            "temperature": rng.normal(37, 0.3, n).round(1),
+            "sao2": (rng.normal(85, 2, n) if low_map else rng.normal(97, 1, n)).round(),
+            "heartrate": rng.normal(85, 5, n).round(),
+            "respiration": rng.normal(18, 2, n).round(),
+            "cvp": np.nan,
+            "etco2": np.nan,
+            "systemicsystolic": rng.normal(120, 8, n).round(),
+            "systemicdiastolic": rng.normal(70, 5, n).round(),
+            "systemicmean": (rng.normal(60, 3, n) if low_map else rng.normal(85, 5, n)).round(),
+            "pasystolic": np.nan,
+            "padiastolic": np.nan,
+            "pamean": np.nan,
+            "st1": np.nan,
+            "st2": np.nan,
+            "st3": np.nan,
+            "icp": np.nan,
+        }
+    )
+
+
[…]
