### 2026-06-09 1059116
chore: replace loguru with stdlib logging, drop loguru dependency, refresh lockfile
