+++ clinops/preprocess/units.py
@@ -27,12 +27,14 @@ Supported conversion families
 
 from __future__ import annotations
 
+import logging
 from collections.abc import Callable
 from dataclasses import dataclass
 from typing import Any
 
 import pandas as pd
-from loguru import logger
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass
