+++ clinops/orchestrate/stepfunctions.py
@@ -18,11 +18,12 @@ Requires ``pip install clinops[aws]``.
 from __future__ import annotations
 
 import json
+import logging
 import time
 from dataclasses import dataclass, field
 from typing import Any
 
-from loguru import logger
+logger = logging.getLogger(__name__)
 
 
 @dataclass
