+++ clinops/preprocess/outliers.py
@@ -23,12 +23,14 @@ Circulation, 101(23), e215–e220.
 
 from __future__ import annotations
 
+import logging
 from dataclasses import dataclass
 from typing import Any
 
 import numpy as np
 import pandas as pd
-from loguru import logger
+
+logger = logging.getLogger(__name__)
 
 
 @dataclass
