+++ pyproject.toml
@@ -32,7 +32,6 @@ dependencies = [
     "pydantic>=2.5",
     "scipy>=1.11",
     "scikit-learn>=1.4",
-    "loguru>=0.7",
 ]
 
 [project.optional-dependencies]
