### 2026-06-09 d773fe0
release v0.4.0
