### 2026-06-09 c61d685
fix(eicu): address PR #8 review comments + config/splitter hardening; standardize on ruff format
