+++ clinops/ingest/eicu.py
@@ -62,7 +62,7 @@ from pathlib import Path
 
 import numpy as np
 import pandas as pd
-from pydantic import BaseModel, field_validator
+from pydantic import BaseModel, field_validator, model_validator
 
 from clinops.ingest.mappings.eicu_lab_map import canonical_lab_features, map_labname
 from clinops.ingest.schema import LeakageError, SchemaValidationError
@@ -343,9 +343,15 @@ class EicuLoader:
         usecols = None
         if columns is not None:
             wanted = {"patientunitstayid", "observationoffset", *columns}
-            # Resolve against header so we only request columns that exist.
-            header = pd.read_csv(path, nrows=0)
-            usecols = [c for c in header.columns if c in wanted]
+            # Resolve against the actual file schema so we only request columns
+            # that exist — and never hand a Parquet file to the CSV reader.
+            if path.suffix == ".parquet":
+                import pyarrow.parquet as pq
+
+                available: list[str] = pq.read_schema(path).names
+            else:
+                available = list(pd.read_csv(path, nrows=0).columns)
+            usecols = [c for c in available if c in wanted]
 
         logger.debug(
             "Loading eICU %s from %s (chunked, chunk_size=%d)",
@@ -548,9 +554,17 @@ class EicuCohortConfig(BaseModel):
         Length of the input window. Default 24.
     prediction_hours:
         Length of the forward label horizon. Default 6.
+    baseline_hours:
+        Window (from ICU admission) used to compute each patient's physiological
+        baseline (lowest creatinine, best GCS) for the renal and neurological
+        labels. Default 6. Kept separate from ``prediction_hours`` — they are
+        unrelated concepts that merely share a default.
     missingness_threshold:
         Stays whose critical-variable missingness on the hourly grid exceeds
         this fraction are excluded. Default 0.30.
+    dnr_window_hours:
+        Stays with a DNR / care-limitation order within this many hours of
+        admission are excluded. Default 6.
     first_stay_only:
         Keep only the first ICU stay per ``uniquepid``. Default ``True``.
     hospital_ids:
@@ -568,12 +582,39 @@ class EicuCohortConfig(BaseModel):
     max_icu_hours: int = 72
     observation_hours: int = 24
     prediction_hours: int = 6
+    baseline_hours: int = 6
     missingness_threshold: float = 0.30
     dnr_window_hours: int = 6
     first_stay_only: bool = True
     hospital_ids: list[int] | None = None
     limit_stays: int | None = None
     random_state: int = 42
+
+    @model_validator(mode="after")
+    def _check_consistency(self) -> EicuCohortConfig:
+        positive = {
+            "min_age": self.min_age,
+            "min_los_hours": self.min_los_hours,
+            "max_icu_hours": self.max_icu_hours,
+            "observation_hours": self.observation_hours,
+            "prediction_hours": self.prediction_hours,
+            "baseline_hours": self.baseline_hours,
+            "dnr_window_hours": self.dnr_window_hours,
+        }
+        for name, value in positive.items():
+            if value <= 0:
+                raise ValueError(f"{name} must be positive, got {value}")
+        if not 0.0 <= self.missingness_threshold <= 1.0:
+            raise ValueError(
+                f"missingness_threshold must be in [0, 1], got {self.missingness_threshold}"
+            )
+        if self.observation_hours + self.prediction_hours > self.max_icu_hours:
+            raise ValueError(
+                f"observation_hours + prediction_hours "
[…]
