+++ pyproject.toml
@@ -40,7 +40,7 @@ gcp  = ["google-cloud-storage>=2.13", "google-cloud-bigquery>=3.14"]
 aws  = ["boto3>=1.34"]
 dev  = [
     "pytest>=7.4", "pytest-cov>=4.1", "pytest-mock>=3.12",
-    "black==26.1.0", "ruff==0.15.2", "mypy==1.19.1", "pandas-stubs>=2.1",
+    "ruff==0.15.2", "mypy==1.19.1", "pandas-stubs>=2.1",
     "pydantic>=2.5",
     "mkdocs>=1.5", "mkdocs-material>=9.5", "mkdocstrings[python]>=0.24",
     "pre-commit>=3.7",
@@ -54,10 +54,6 @@ Issues        = "https://github.com/chaitanyakasaraneni/clinops/issues"
 
 [tool.hatch.build.targets.wheel]
 packages = ["clinops"]
-
-[tool.black]
-line-length    = 100
-target-version = ["py312"]
 
 [tool.ruff]
 line-length    = 100

[…]
