### 2026-06-09 1bca8a2
fix(eicu): typed pyarrow schema read, validated build_sequences overrides, hardened DNR guard
- _load_large: suppress no-untyped-call on pyarrow.parquet.read_schema (CI mypy --strict)
- build_sequences: route per-call overrides through EicuCohortConfig validation
  (rejects non-positive values and observation+prediction > max_icu_hours)
- _dnr_stays: include patientunitstayid in the missing-column guard (clean skip, no KeyError)
- tests: invalid-override rejection + malformed carePlanGeneral skip
